Fix issues with ios-click
".ios-click" was only working with the first instance of ".ios-click" in the file in the previous commit. This
commit fixes that while also fixing animation issue on the arrow (animates properly now).
"buttonHover" has not been added to the classes and next task will be creating a function that adds
or removes "buttonHover" class based on the display size of window.